requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-environ==0.6.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-registration-redux==2.9
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
idna==3.2
Pillow==8.3.2
pip==21.2.4
python-decouple==3.4
pytz==2021.1
requests==2.26.0
setuptools==57.4.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
